requirements.txt
pytest==8.0.0
Faker==22.6.0
requests==2.31.0 
allure-python-commons~=2.13.2